rename LV_X64 to LV_WIN64
git-svn-id: svn://23.27.117.66/h2/trunk@236 54bad739-14b1-1b4c-9595-0bb2987a2747

+++ lv/src/lv/Console.cpp
@@ -32,7 +32,7 @@ namespace lv
 		}
 
 		SetConsoleTitleA(title);
-		long lStdHandle = (long)GetStdHandle(STD_ERROR_HANDLE);
+		intptr_t lStdHandle = (intptr_t)GetStdHandle(STD_ERROR_HANDLE);
 		int hConHandle = _open_osfhandle(lStdHandle, _O_TEXT);       
 		FILE* fp = _fdopen(hConHandle,"w");
 		if (fp == nullptr)
